+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [1.5.0] - Unreleased
 ### Added
 - New plugin `esbuild`.
+- New plugin `parcel_css`.
 - New upgrade notification if it's detected an old version of Lume.
 - New architecture for the local server, with support for middlewares.
 - New command `lume import-map` to create/update import maps files [#164].